+++ src/layouts/DappSubmissionLayout/containers/SummaryContainer/SummaryContainer.vue
@@ -1,6 +1,6 @@
-<template>
+<template functional>
   <div class="mt-5 ">
-    <p v-if="lackOfInfo" class="error">
+    <p v-if="props.lackOfInfo" class="error">
       Please provide all the required information we need.
     </p>
     <b-container class="dapp-info-container mx-auto mt-2 pb-5 pl-5 w-75">
@@ -9,12 +9,12 @@
         <b-col>
           <div class="mt-5">
             <p class="summary-title mb-3">Dapp name</p>
-            <p class="summary-content">{{ form.dappName }}</p>
+            <p class="summary-content">{{ props.form.dappName }}</p>
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3 mb-3">Tags</p>
             <div
-              v-for="(tag, idx) in form.tags"
+              v-for="(tag, idx) in props.form.tags"
               :key="tag + idx"
               class="summary-tag-content mr-2 mb-2"
             >
@@ -25,56 +25,62 @@
             <p class="summary-title mb-3">Mock user flow</p>
             <p class="summary-content">
               <img
-                v-if="form.mockFlowFile.type !== 'application/pdf'"
-                :src="form.mockFlowUrl"
+                v-if="props.form.mockFlowFile.type !== 'application/pdf'"
+                :src="props.form.mockFlowUrl"
                 class="summary-img"
               />
-              <span class="align-middle"> {{ form.mockFlowFile.name }} </span>
+              <span class="align-middle">
+                {{ props.form.mockFlowFile.name }}
+              </span>
             </p>
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3">Contract address</p>
-            <p class="summary-content">{{ form.contractAddress }}</p>
+            <p class="summary-content">{{ props.form.contractAddress }}</p>
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3">Description</p>
-            <span class="summary-content">{{ form.description }}</span>
+            <span class="summary-content">{{ props.form.description }}</span>
           </div>
         </b-col>
         <b-col class="pl-5">
           <div class="mt-5">
             <p class="summary-title mb-3">Category</p>
-            <p class="summary-content">{{ form.category }}</p>
+            <p class="summary-content">{{ props.form.category }}</p>
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3">Launch in the U.S. market</p>
-            <p class="summary-content">{{ form.usMarket }}</p>
+            <p class="summary-content">{{ props.form.usMarket }}</p>
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3">Dapp icon</p>
             <p class="summary-content">
-              <img :src="form.dappIconUrl" class="summary-img" />
-              <span class="align-middle"> {{ form.dappIconFile.name }} </span>
+              <img :src="props.form.dappIconUrl" class="summary-img" />
+              <span class="align-middle">
+                {{ props.form.dappIconFile.name }}
+              </span>
             </p>
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3">Banner image</p>
             <p class="summary-content">
-              <img :src="form.bannerUrl" class="summary-img" />
-              <span class="align-middle"> {{ form.bannerFile.name }} </span>
+              <img :src="props.form.bannerUrl" class="summary-img" />
+              <span class="align-middle">
+                {{ props.form.bannerFile.name }}
+              </span>
             </p>
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3">Dapp website</p>
-            <p class="summary-content">{{ form.dappWebsite }}</p>
+            <p class="summary-content">{{ props.form.dappWebsite }}</p>
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3">Dapp contract audit</p>
-            <p class="summary-content">{{ form.contractAudit }}</p>
+            <p class="summary-content">{{ props.form.contractAudit }}</p>
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3">Dapp status</p>
-            <p class="summary-content">{{ form.dappStatus }}</p>
+            <p class="summary-content">{{ props.form.dappStatus }}</p>
           </div>
         </b-col>
       </b-row>
@@ -85,26 +91,26 @@
         <b-col>
           <div class="mt-5">
             <p class="summary-title mb-3">Authors</p>
-            <p class="summary-content">{{ form.authors }}</p>
+            <p class="summary-content">{{ props.form.authors }}</p>
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3 mb-3">My full name</p>
-            <p class="summary-content">{{ form.fullName }}</p>
+            <p class="summary-content">{{ props.form.fullName }}</p>
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3">My email address</p>
-            <p class="summary-content">{{ form.email }}</p>
+            <p class="summary-content">{{ props.form.email }}</p>
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3">Company website</p>
-            <p class="summary-content">{{ form.companyWebsite }}</p>
+            <p class="summary-content">{{ props.form.companyWebsite }}</p>
           </div>
         </b-col>
         <b-col class="pl-5">
           <div class="mt-5">
             <p class="summary-title mb-3">Social media</p>
             <div
-              v-for="(acct, idx) in socialAccts"
+              v-for="(acct, idx) in props.socialAccts"
               :key="acct + idx"
               class="social-links mr-2"
             >
@@ -113,11 +119,11 @@
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3">Software License</p>
-            <p class="summary-content">{{ form.license }}</p>
+            <p class="summary-content">{{ props.form.license }}</p>
           </div>
           <div class="mt-5">
             <p class="summary-title mb-3">Additional notes</p>
-            <p class="summary-content">{{ form.additionalNotes }}</p>
+            <p class="summary-content">{{ props.form.additionalNotes }}</p>
           </div>
         </b-col>
       </b-row>