+++ src/components/DecisionTree/DecisionTree.vue
@@ -1,7 +1,8 @@
 <template>
-  <div class="">
+  <div class="decision-tree-container">
     <button :class="button ? 'active' : ''" class="show-button" @click="toggle">
       <img src="@/assets/images/icons/DecisionTree/need-help.svg" />
+      <p>Help</p>
     </button>
 
     <customer-support :no-icon="true" :show="showCustomerSupport" />