+++ src/layouts/CreateWalletLayout/components/TutorialModal/TutorialModal.vue
@@ -34,9 +34,6 @@ export default {
       type: Function,
       default: function() {}
     }
-  },
-  data() {
-    return {};
   }
 };
 </script>